requirements.txt
# AI Expert - Required Dependencies

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# UI Framework
streamlit>=1.28.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Excel Support
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Advanced ML (Optional but recommended)
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Model Persistence
joblib>=1.3.0

# Report Generation
reportlab>=4.0.0
python-docx>=1.0.0

# AI Integration
groq>=0.4.0
google-generativeai>=0.3.0
requests>=2.31.0

# Additional Data Formats
pyarrow>=14.0.0
sqlalchemy>=2.0.0

# EXE Building (Optional - for development only)
pyinstaller>=6.0.0
